+++ src/routes/ConsortiumManager/utils/index.js
@@ -1,3 +1,5 @@
+export { filterSectionsByInterfaces } from './filterSectionsByInterfaces';
 export { groupMembersSettings } from './groupMembersSettings';
+export { isModuleInterfacesAvailable } from './isModuleInterfacesAvailable';
 export { validateNameAndCodeRequired } from './validateNameAndCodeRequired';
 export { validateNameRequired } from './validateNameRequired';